backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
Flask-Cors==4.0.0
WeasyPrint==60.2
gunicorn==21.2.0